Refactor routes into subpackages
