+++ app/api.py
@@ -5,8 +5,6 @@ from app.routes import (
     shop_bp,
     item_bp,
     consumer_bp,
-    wallet_bp,
-    order_bp,
     admin_bp,
     agent_bp,
 )
@@ -22,8 +20,6 @@ def register_api_v1(app):
     app.register_blueprint(shop_bp)
     app.register_blueprint(item_bp)
     app.register_blueprint(consumer_bp)
-    app.register_blueprint(wallet_bp)
-    app.register_blueprint(order_bp)
     app.register_blueprint(admin_bp)
     if os.environ.get("OPENAI_API_KEY") and agent_bp:
         logging.info("Agent blueprint enabled")
